refactor(Favorite): изменена файловая архитектура
