+++ EatHub/EatHub/Modules/Favorite/RecipeRow.swift
@@ -1,0 +1,34 @@
+import SwiftUI
+
+struct RecipeRow: View {
+    let recipe: Recipe
+    let onToggleFavorite: () -> Void
+
+    var body: some View {
+        HStack {
+            Image(recipe.imageName)
+                .resizable()
+                .aspectRatio(contentMode: .fill)
+                .frame(width: 60, height: 60)
+                .clipped()
+                .cornerRadius(8)
+
+            Text(recipe.name)
+                .font(.headline)
+                .padding(.leading, 8)
+
+            Spacer()
+
+            Button(action: onToggleFavorite) {
+                Image(systemName: recipe.isFavorite ? "heart.fill" : "heart")
+                    .foregroundColor(recipe.isFavorite ? .red : .gray)
+                    .font(.system(size: 22))
+            }
+        }
+        .padding()
+        .foregroundColor(.primary)
+        .background(Color(.systemBackground))
+        .cornerRadius(12)
+        .shadow(color: Color.black.opacity(0.05), radius: 2, x: 0, y: 1)
+    }
+}
