+++ EatHub/EatHub/Modules/Models/Recipe.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct Recipe: Identifiable, Equatable {
+    let id = UUID()
+    var name: String
+    var imageName: String
+    var isFavorite: Bool = true
+}
